% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_are_connected.R
\name{dru_are_connected}
\alias{dru_are_connected}
\title{Check if the two addresses are connected.}
\usage{
dru_are_connected(connection, start, end, address1, address2, directed)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}

\item{address1}{A string that contains the first address to be checked.}

\item{address2}{A string that contains the second address to be checked.}

\item{directed}{A boolean that specifies if the graph should be considered
directed or undirected.}
}
\value{
A data frame, containing all parameters
as well as the boolean are_connected
}
\description{
The function sends a request to a Dru instance
and checks if two given addresses are connected.
}
\details{
Returns true/false information whether two addresses
are connected within a given range of blocks. \cr
If any of these addresses does not exist in the graph, None will be returned.
}
\examples{
dru_are_connected("https://dru.bergplace.org/api", 0, 100,
"t1StbPM4X3j4FGM57HpGnb9BMbS7C1nFW1r", "t1KstPVzcNEK4ZeauQ6cogoqxQBMDSiRnGr", "true")
}
\author{
Mateusz Gabrys
}
