% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_blocks.R
\name{dru_get_blocks}
\alias{dru_get_blocks}
\title{Return the list of blocks}
\usage{
dru_get_blocks(connection, start, end)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}
}
\value{
This function returns a JSON containing detailed information
about blocks in a given range.
}
\description{
The function sends a request to a Dru instance
and returns the list of blocks for a given range of block's height.
}
\details{
Returns the list of blocks for a given range of block's heights. \cr
The returned JSON can be used for further processing
if none of the available functions is suitable
for performing the requested analysis. \cr
This function returns whole blocks. No attributes are stripped.
}
\examples{
dru_get_blocks("https://dru.bergplace.org/api", 0, 10)
}
\author{
Mateusz Gabrys
}
