% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_blocks_reduced.R
\name{dru_get_blocks_reduced}
\alias{dru_get_blocks_reduced}
\title{Return the list of blocks with limited fields}
\usage{
dru_get_blocks_reduced(connection, start, end)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}
}
\value{
This function returns a JSON with the following information
about blocks:
\itemize{
\item height
\item time
\item transaction id
\item input addresses
\item output addresses
\item value
}
}
\description{
The function sends a request to a Dru instance
and returns the list of blocks with limited fields
for a given range of block's height.
}
\details{
Returns the list of blocks with limited fields
for a given range of block's heights. \cr
The returned JSON can be used for further processing
if none of the available functions is suitable
for performing the requested analysis.
}
\examples{
dru_get_blocks_reduced("https://dru.bergplace.org/api", 0, 10)
}
\author{
Mateusz Gabrys
}
