% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_degree.R
\name{dru_get_degree}
\alias{dru_get_degree}
\title{Return the value of the degree}
\usage{
dru_get_degree(connection, start, end, mode)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}

\item{mode}{A string that specifies the method to calculate the degree.
Three modes are available : \cr
all - total degree of the node,
in - in degree of the node, out - out degree of the node.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item address
\item degree
}
}
\description{
The function sends a request to a Dru instance and returns
the list of addresses and degree values corresponding to them.
}
\details{
Returns the list of addresses and the value of degree corresponding to them.
\cr
The graph is created from the blocks in the range: start - end. \cr
The graph will be built as directed.
Different measure variants can be computed: all, in, out.
}
\examples{
dru_get_degree("https://dru.bergplace.org/api", 0, 10, "all")
}
\author{
Mateusz Gabrys
}
