% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_degree_max.R
\name{dru_get_degree_max}
\alias{dru_get_degree_max}
\title{Return the value of the degree}
\usage{
dru_get_degree_max(connection, start, end, mode)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}

\item{mode}{A string that specifies the method to calculate the degree.
3 modes are available : \cr all - total degree of the node.
in - in degree of the node, out - out degree of the node.}
}
\value{
This function returns a list including an address, mode
as well as all of the degrees with their heights.
}
\description{
The function sends a request to a Dru instance
and returns addresses with the highest degree value.
}
\details{
Returns an address and the value of the degree corresponding to it. \cr
The graph is created from the blocks in the range: start - end. \cr
The graph will be built as directed,
but using the mode,
all variants of the measure can be computed: all, in, out.
}
\examples{
dru_get_degree_max("https://dru.bergplace.org/api", 0, 1000, "all")
}
\author{
Mateusz Gabrys
}
