% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_density.R
\name{dru_get_density}
\alias{dru_get_density}
\title{Return the density of a graph.}
\usage{
dru_get_density(connection, start, end, directed, loops)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}

\item{directed}{A boolean that specifies if the graph should be considered
directed or undirected.}

\item{loops}{A boolean that specifies if loops should be in the build graph.}
}
\value{
This function returns a number that is the density of the graph.
}
\description{
The function sends a request to a Dru instance
and returns the density of the created graph.
}
\details{
Returns the density of the graph created from the blocks
in the range: start - end. \cr
The graph can be considered as directed or undirected.
}
\examples{
dru_get_density("https://dru.bergplace.org/api", 0, 1000, "true", "true")
}
\author{
Mateusz Gabrys
}
