% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_diameter.R
\name{dru_get_diameter}
\alias{dru_get_diameter}
\title{Return the diameter of a graph.}
\usage{
dru_get_diameter(connection, start, end, directed)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}

\item{directed}{A boolean that specifies if the graph should be considered
directed or undirected.}
}
\value{
This function returns a number that is the diameter of the graph.
}
\description{
The function sends a request to a Dru instance
and returns the diameter of the created graph.
}
\details{
Returns the diameter of the graph created from the blocks
in the range: start - end. \cr
The graph can be considered as directed or undirected.
}
\examples{
dru_get_diameter("https://dru.bergplace.org/api", 0, 1000, "true")
}
\author{
Mateusz Gabrys
}
