% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_transactions_value.R
\name{dru_get_transactions_value}
\alias{dru_get_transactions_value}
\title{Return the count and transaction value between addresses.}
\usage{
dru_get_transactions_value(connection, start, end, address1, address2)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}

\item{address1}{A string that contains the first address to be checked.}

\item{address2}{A string that contains the second address to be checked.}
}
\value{
A data frame, containing all parameters
as well as the transaction_count and transaction_value fields.
}
\description{
The function sends a request to a Dru instance
and returns the count and transaction value between addresses.
}
\details{
Returns the count and total value of transactions
between two addresses in the graph. \cr
The graph is created from the blocks in the range: start - end. \cr
If any of these addresses does not exist in the graph, None will be returned.
}
\examples{
dru_get_transactions_value("https://dru.bergplace.org/api", 0, 100,
"t1StbPM4X3j4FGM57HpGnb9BMbS7C1nFW1r", "t1KstPVzcNEK4ZeauQ6cogoqxQBMDSiRnGr")
}
\author{
Mateusz Gabrys
}
