% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_transitivity.R
\name{dru_get_transitivity}
\alias{dru_get_transitivity}
\title{Return the global transitivity value for a collection of addresses.}
\usage{
dru_get_transitivity(connection, start, end)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item address
\item transitivity
}
}
\description{
The function sends a request to a Dru instance
and returns the value of transitivity for addresses in a created graph.
}
\details{
Returns the nodes’ clustering coefficient in the graph. \cr
The graph is created from the blocks in the range:start - end. \cr
For the global clustering coefficient value,
use the dru_get_transitivity_global() function.
}
\examples{
dru_get_transitivity("https://dru.bergplace.org/api", 0, 1000)
}
\author{
Mateusz Gabrys
}
