% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_get_zcash_tx_types_count.R
\name{dru_get_zcash_tx_types_count}
\alias{dru_get_zcash_tx_types_count}
\title{Return transaction types in a given graph.}
\usage{
dru_get_zcash_tx_types_count(connection, start, end)
}
\arguments{
\item{connection}{A string that contains an address of the Dru instance.}

\item{start}{A number which is the start height of a block
that is included in the search range.}

\item{end}{A number which is the end height of a block
that is included in the search range.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item block_height
\item s.s (shielded-shielded transactions)
\item s.t (shielded-transparent transactions)
\item t.s (transparent-shielded transactions)
\item t.t (transparent-transparent transactions)
\item time
\item date_time
}
}
\description{
The function sends a request to a Dru instance
and returns block's height and the types of transactions.
}
\details{
The returned data frame contains the blocks heights accompanied
by the following: \cr block time and number of shielded-shielded, shielded-transparent,
transparent-shielded and transparent-transparent transactions. \cr
For details on transactions types, see: https://z.cash/technology/
}
\examples{
dru_get_zcash_tx_types_count("https://dru.bergplace.org/api", 0, 100)
}
\author{
Mateusz Gabrys
}
