% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dru_start.R
\name{dru_start}
\alias{dru_start}
\title{Initialize Dru connection.}
\usage{
dru_start(address, https = TRUE)
}
\arguments{
\item{address}{A string that contains an address of the Dru instance}

\item{https}{A boolean that specifies, if we want to use https connection}
}
\value{
String with either http or https address of the instance
}
\description{
Function connects to a Dru instance, and checks if the connections work.
}
\details{
This function has been designed to make working with the RDru package easier.
It sets up the connection between the user and the Dru instance.
To use the function we have to provide the address of the instance as well as
specify, whether we want to use https or not.
It is recommended to run the dru_start function
and assign it's result to a variable,
before running other functions included in the RDru package.
}
\examples{
dru_start("dru.bergplace.org/api", TRUE)
}
\author{
Mateusz Gabrys
}
