% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RDru}
\alias{RDru}
\alias{RDru-package}
\title{RDru: Analyze graphs based on transactions in blockchain}
\description{
Easily use the endpoints available on the Dru website :
https://dru.readthedocs.io/en/master/docs/api-docs.html
using R environment.
}
\author{
\strong{Maintainer}: Mateusz Gabrys \email{mgabrys777@gmail.com}

}
